Renamed sys_mmap.* to smmap.*
