+++ sys/mmap.c
@@ -1,4 +1,4 @@
-#include <sys_mmap.h>
+#include <mmap.h>
 #include <thread.h>
 #include <stdc.h>
 #include <errno.h>
